+++ index.ts
@@ -1,4 +1,5 @@
 import './src/operators';
+
 export * from './src/angular-cesium.module';
 export * from './src/models';
 export * from './src/components';